+++ Aliucord/src/main/java/com/aliucord/coreplugins/slashcommandsfix/RemoteApplicationCommand.java
@@ -1,0 +1,19 @@
+/*
+ * This file is part of Aliucord, an Android Discord client mod.
+ * Copyright (c) 2021 Juby210 & Vendicated
+ * Licensed under the Open Software License version 3.0
+ */
+
+package com.aliucord.coreplugins.slashcommandsfix;
+
+import com.discord.models.commands.ApplicationCommandOption;
+import java.util.List;
+
+class RemoteApplicationCommand extends com.discord.models.commands.RemoteApplicationCommand {
+    public Permissions permissions_;
+
+    public RemoteApplicationCommand(String id, long applicationId, String name, String description, List<ApplicationCommandOption> options, Permissions permissions, String version) {
+        super(id, applicationId, name, description, options, null, version, true, null, null); // TODO: defaultPermissions
+        this.permissions_ = permissions;
+    }
+}
