+++ Aliucord/src/main/java/com/aliucord/coreplugins/slashcommandsfix/RequestSource.java
@@ -1,0 +1,13 @@
+/*
+ * This file is part of Aliucord, an Android Discord client mod.
+ * Copyright (c) 2021 Juby210 & Vendicated
+ * Licensed under the Open Software License version 3.0
+ */
+
+package com.aliucord.coreplugins.slashcommandsfix;
+
+enum RequestSource {
+    GUILD,
+    BROWSE,
+    QUERY;
+}
